Entrance Super Refactor, Ajax, many things, new things to model, awesome coding refactor, such much coding

+++ db/migrate/20150503195621_destroy_order_scaffold_order.rb
@@ -1,6 +1,8 @@
 class DestroyOrderScaffoldOrder < ActiveRecord::Migration
 	# In order to scaffold de basic order views, Im destroying the table to later run "rails g scaffold order"
   def change
-  	drop_table :orders
+  	if table_exists? :orders
+  		drop_table :orders
+  	end
   end
 end
